+++ src/phpeggy.js
@@ -1,25 +1,31 @@
 "use strict";
 
+const path = require("path");
+
 exports.use = function(config, options) {
-  config.passes.check.push(
-    require("./passes/report-mbstring-incompatibility")
-  );
   config.passes.generate = [
     require("./passes/generate-bytecode-php"),
     require("./passes/generate-php"),
   ];
-  options.output = "source";
   if (!options.phpeggy) {
     options.phpeggy = {};
   }
+
+  // Before peggy 4.1 it was possible to use the option output=source
+  // and there was no difference between the cli and the js api.
+  // However, the cli relies on source.code from the ast compare to
+  // https://github.com/peggyjs/peggy/blob/7a8fd86d43ff14ddfd2f1f5034c47bb6db380cc9/bin/peggy-cli.js#L517)
+  // We use this dirty hack to detect if we are in the cli or not.
+  if (!new Error().stack.includes(path.join("peggy", "bin", "peggy-cli.js"))) {
+    options.output = "source";
+  }
+
   if (options.phpeggy.parserNamespace === undefined) {
     options.phpeggy.parserNamespace = "PHPeggy";
   }
+
   if (options.phpeggy.parserClassName === undefined) {
     options.phpeggy.parserClassName = "Parser";
-  }
-  if (options.phpeggy.mbstringAllowed === undefined) {
-    options.phpeggy.mbstringAllowed = true;
   }
 };
 /*